+++ OpenGL-Learning/src/Game/world/Chunk.h
@@ -65,7 +65,7 @@ public:
 	void OnUpdate();
 
 	// Members
-	inline const unsigned int getAmountBlockStatic() const { return m_BlockStatic.size(); }
+	inline const size_t getAmountBlockStatic() const { return m_BlockStatic.size(); }
 	inline const unsigned int getDrawCalls() {
 		const unsigned int calls = m_DrawCalls;
 		m_DrawCalls = 0;
